fix S3 List* pagination CommonPrefixes
